+++ eras/conway/impl/src/Cardano/Ledger/Conway/Rules/Deleg.hs
@@ -31,7 +31,7 @@ import Cardano.Ledger.Binary.Coders (
   (!>),
   (<!),
  )
-import Cardano.Ledger.CertState (CertState (..), DState (..))
+import Cardano.Ledger.CertState (CertState (..), DState (..), vsDReps)
 import Cardano.Ledger.Coin (Coin)
 import Cardano.Ledger.Conway.Core
 import Cardano.Ledger.Conway.Era (ConwayDELEG, ConwayEra)
@@ -40,12 +40,13 @@ import Cardano.Ledger.Conway.TxCert (
   Delegatee (DelegStake, DelegStakeVote, DelegVote),
  )
 import Cardano.Ledger.Credential (Credential)
+import Cardano.Ledger.DRep (DRep (..))
 import Cardano.Ledger.Keys (KeyHash, KeyRole (..))
 import Cardano.Ledger.PoolParams (PoolParams)
 import qualified Cardano.Ledger.Shelley.HardForks as HF
 import qualified Cardano.Ledger.UMap as UM
 import Control.DeepSeq (NFData)
-import Control.Monad (forM_, guard)
+import Control.Monad (forM_, guard, unless)
 import Control.State.Transition (
   BaseM,
   Environment,
@@ -71,11 +72,7 @@ import NoThunks.Class (NoThunks)
 
 data ConwayDelegEnv era = ConwayDelegEnv
   { cdePParams :: PParams era
-  , cdePools ::
-      !( Map
-          (KeyHash 'StakePool (EraCrypto era))
-          (PoolParams (EraCrypto era))
-       )
+  , cdePools :: Map (KeyHash 'StakePool (EraCrypto era)) (PoolParams (EraCrypto era))
   }
   deriving (Generic)
 
@@ -87,18 +84,19 @@ instance EraPParams era => EncCBOR (ConwayDelegEnv era) where
             !> To cdePParams
             !> To cdePools
 
-instance NFData (PParams era) => NFData (ConwayDelegEnv era)
+instance (Era era, NFData (PParams era)) => NFData (ConwayDelegEnv era)
 
 deriving instance Eq (PParams era) => Eq (ConwayDelegEnv era)
 
 deriving instance Show (PParams era) => Show (ConwayDelegEnv era)
 
 data ConwayDelegPredFailure era
-  = IncorrectDepositDELEG !Coin
-  | StakeKeyRegisteredDELEG !(Credential 'Staking (EraCrypto era))
-  | StakeKeyNotRegisteredDELEG !(Credential 'Staking (EraCrypto era))
-  | StakeKeyHasNonZeroRewardAccountBalanceDELEG !Coin
-  | DelegateeNotRegisteredDELEG !(KeyHash 'StakePool (EraCrypto era))
+  = IncorrectDepositDELEG Coin
+  | StakeKeyRegisteredDELEG (Credential 'Staking (EraCrypto era))
+  | StakeKeyNotRegisteredDELEG (Credential 'Staking (EraCrypto era))
+  | StakeKeyHasNonZeroRewardAccountBalanceDELEG Coin
+  | DelegateeDRepNotRegisteredDELEG (Credential 'DRepRole (EraCrypto era))
+  | DelegateeNotRegisteredDELEG (KeyHash 'StakePool (EraCrypto era))
   deriving (Show, Eq, Generic)
 
 type instance EraRuleFailure "DELEG" (ConwayEra c) = ConwayDelegPredFailure (ConwayEra c)
@@ -122,6 +120,8 @@ instance Era era => EncCBOR (ConwayDelegPredFailure era) where
         Sum (StakeKeyNotRegisteredDELEG @era) 3 !> To stakeCred
       StakeKeyHasNonZeroRewardAccountBalanceDELEG mCoin ->
         Sum (StakeKeyHasNonZeroRewardAccountBalanceDELEG @era) 4 !> To mCoin
+      DelegateeDRepNotRegisteredDELEG delegatee ->
+        Sum (DelegateeDRepNotRegisteredDELEG @era) 5 !> To delegatee
       DelegateeNotRegisteredDELEG delegatee ->
         Sum (DelegateeNotRegisteredDELEG @era) 6 !> To delegatee
 
@@ -131,6 +131,7 @@ instance Era era => DecCBOR (ConwayDelegPredFailure era) where
     2 -> SumD StakeKeyRegisteredDELEG <! From
     3 -> SumD StakeKeyNotRegisteredDELEG <! From
     4 -> SumD StakeKeyHasNonZeroRewardAccountBalanceDELEG <! From
+    5 -> SumD DelegateeDRepNotRegisteredDELEG <! From
     6 -> SumD DelegateeNotRegisteredDELEG <! From
     n -> Invalid n
 
@@ -183,10 +184,18 @@ conwayDelegTransition = do
     checkStakeDelegateeRegistered =
       let checkPoolRegistered targetPool =
             targetPool `Map.member` pools ?! DelegateeNotRegisteredDELEG targetPool
+          checkDRepRegistered = \case
+            DRepAlwaysAbstain -> pure ()
+            DRepAlwaysNoConfidence -> pure ()
+            DRepCredential targetDRep -> do
+              let dReps = vsDReps (certVState certState)
+              unless (HF.bootstrapPhase (pp ^. ppProtocolVersionL)) $
+                targetDRep `Map.member` dReps ?! DelegateeDRepNotRegisteredDELEG targetDRep
        in \case
             DelegStake targetPool -> checkPoolRegistered targetPool
-            DelegStakeVote targetPool _ -> checkPoolRegistered targetPool
-            DelegVote _ -> pure ()
+            DelegStakeVote targetPool targetDRep ->
+              checkPoolRegistered targetPool >> checkDRepRegistered targetDRep
+            DelegVote targetDRep -> checkDRepRegistered targetDRep
   case cert of
     ConwayRegCert stakeCred sMayDeposit -> do
       forM_ sMayDeposit checkDepositAgainstPParams