+++ eras/conway/impl/CHANGELOG.md
@@ -36,6 +36,7 @@
 * Add a new field to `GovInfoEvent` and change "unclaimed" field from `Set` to a `Map`.
 * Changed return type of `proposalsShowDebug`
 * Added `gen-golden` executable needed for golden tests: #4629
+* Add `DelegateeDRepNotRegisteredDELEG` predicate failure
 
 ### `testlib`